WIP: refactoring and code cleanup - move utility stuff into util/ subfolder - split up util materials into separate files - make everything in mc lib in mc:: namespace, with utils in mc::util:: namespace
