+++ .vscode/tasks.json
@@ -34,7 +34,7 @@
                 "focus": false,
                 "panel": "shared",
                 "showReuseMessage": true,
-                "clear": false
+                "clear": true
             }
         }
     ]
